[BSE-4011] Refactor bcast to use C++ implementation (#8367)
* Rename bcast to bcast_preallocated

* replace one more use

* remove unnecessary bcast import

* use C++ broadcast for arrays

* avoid leak

* rename bcast_comm to bcast

* minor f-string refactor

* support table type

* fix RangeIndex bug

* minor cleanup

* add comm ranks to broadcast_array

* handle ref_arr nullptr case

* fix comm_ranks return

* fix dataframe case

* make comm_ranks optional

* free MPI communicator

* remove unnecessary copy

* remove comm from c_bcast

* fix merge conflict and add unittest [run ci]

* [run CI]

* add no cover pragmas [run CI]

* minor comment fix

* fix build error

* fix assert error

* update docstrings

* [run CI]

+++ bodo/libs/_bodo_common.cpp
@@ -532,7 +532,7 @@ std::unique_ptr<Schema> Schema::Project(
     std::vector<std::unique_ptr<DataType>> dtypes;
     dtypes.reserve(column_indices.size());
     for (int64_t col_idx : column_indices) {
-        assert(col_idx < this->column_types.size());
+        assert((size_t)col_idx < this->column_types.size());
         dtypes.push_back(this->column_types[col_idx]->copy());
     }
     return std::make_unique<Schema>(std::move(dtypes));
